[Feat] 체크한 조건은 상단 필터에 표시
